DABOM-427 feat: 어드민 전역 에러 UI 적용
