+++ apps/admin/src/app/(afterLogin)/policy/@modal/(.)[id]/page.tsx
@@ -1,7 +1,7 @@
-import PolicyDetailModal from 'src/components/policy/PolicyDetail';
+import PolicyDetailDrawer from 'src/components/policy/PolicyDetail';
 
 const PolicyDetailPage = () => {
-  return <PolicyDetailModal />;
+  return <PolicyDetailDrawer />;
 };
 
 export default PolicyDetailPage;
