+++ apps/admin/src/app/(afterLogin)/policy/[id]/page.tsx
@@ -1,9 +1,9 @@
-import PolicyDetailModal from 'src/components/policy/PolicyDetail';
+import PolicyDetailDrawer from 'src/components/policy/PolicyDetail';
 
 export default function PolicyDetailPage() {
   return (
     <div className="p-10">
-      <PolicyDetailModal />
+      <PolicyDetailDrawer />
     </div>
   );
 }
