+++ apps/admin/src/app/(afterLogin)/policy/page.tsx
@@ -7,6 +7,7 @@ import { usePathname, useRouter, useSearchParams } from 'next/navigation';
 import { Table } from '@shared';
 
 import { useGetPolicy } from 'src/api/policy/useGetPolicy';
+import Error from 'src/components/common/Error';
 import Loading from 'src/components/common/Loading';
 import Pagination from 'src/components/common/Pagination';
 import { formatPolicy } from 'src/utils/formatPolicy';
@@ -33,7 +34,7 @@ const PolicyContent = () => {
   };
 
   if (!data || !data.policies || data.policies.length === 0) {
-    return <div>표시할 정책이 없습니다.</div>;
+    return <Error title="정책이 존재하지 않습니다." />;
   }
 
   const policyRows = formatPolicy({ policies: data.policies });
